+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 - Remove getAttributeOptionValueRepository, getFlexibleValueRepository from FlexibleManager
 - Attribute fixtures format has changed
 - Product associations import/export format has changed.
+- Rename Association by AssociationType and all properties/methods linked to this class.
 
 # 1.0.0-beta-3 - "Hare Conditioned"